pyproject.toml
[project]
name = "powerpoint-agent-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonschema>=4.25.1",
    "pandas>=2.3.2",
    "pillow>=12.0.0",
    "python-pptx==0.6.23",
]
